# docs/instant-messaging-api-spec.yml
openapi: 3.0.3
info:
  title: Instant Messaging API
  description: API for managing messaging channels
  version: 0.1.0
  # Antonio Paulino - 50512
  # Bernardo Pereira - 50493
servers:
  - url: http://localhost:8080/api
    description: Local server
...
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: Get a list of all users
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Name'
        - $ref: '#/components/parameters/GetCount'
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [ id, name, email ]
            default: name
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersPaginatedOutputModel'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/users/{userID}:
    get:
      tags:
        - Users
      summary: Get details of a specific user by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserID'
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOutputModel'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/users/{userID}/channels:
    get:
      tags:
        - Users
      summary: Get a list of channels the user is a member of
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserID'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/GetCount'
        - name: after
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
            description: Get channels after this ID
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [ id, name ]
            default: name
      responses:
        '200':
          description: List of channels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelsPaginatedOutputModel'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/channels/{channelID}/messages:
    get:
      tags:
        - Messaging
      summary: Get messages from a channel
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/GetCount'
        - $ref: '#/components/parameters/Sort'
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [ id, createdAt, editedAt ]
            default: createdAt
          description: Sort by field
        - name: before
          in: query
          required: false
          schema:
            type: string
            format: date-time
            description: Get messages created after this time
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessagesPaginatedOutputModel'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'


    post:
      tags:
        - Messaging
      summary: Create a message in a channel
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreationInputModel'
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageCreationOutputModel'
        '400':
            $ref: '#/components/responses/BadRequest'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/channels/{channelID}/messages/{messageID}:
    get:
      tags:
        - Messaging
      summary: Get a message from a channel
      security:
          - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
        - $ref: '#/components/parameters/MessageID'
      responses:
        '200':
          description: Message retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageOutputModel'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    put:
      tags:
        - Messaging
      summary: Edit a message in a channel
      security:
          - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
        - $ref: '#/components/parameters/MessageID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreationInputModel'
      responses:
        '200':
          description: Message edited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageUpdateOutputModel'
        '400':
            $ref: '#/components/responses/BadRequest'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    delete:
      tags:
        - Messaging
      summary: Delete a message from a channel
      security:
          - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ChannelID'
        - $ref: '#/components/parameters/MessageID'
      responses:
        '204':
          description: Message deleted successfully
        '400':
            $ref: '#/components/responses/BadRequest'
        '401':
            $ref: '#/components/responses/Unauthorized'
        '403':
            $ref: '#/components/responses/Forbidden'
        '404':
            $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  schemas:
...
    UserOutputModel:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
...
    UsersPaginatedOutputModel:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserOutputModel'
        pagination:
          $ref: '#/components/schemas/PaginationOutputModel'
...
    PaginationOutputModel:
      type: object
      properties:
        current:
          type: integer
        next:
          type: integer
        previous:
          type: integer
        totalPages:
          type: integer
        total:
          type: integer
...
    ChannelOutputModel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        isPublic:
          type: boolean
        owner:
          $ref: '#/components/schemas/UserOutputModel'
        members:
          type: array
          items:
            $ref: '#/components/schemas/ChannelMemberOutputModel'
        createdAt:
          type: string
          format: date-time
    ChannelMemberOutputModel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
          enum: [ owner, member, guest ]
    ChannelsPaginatedOutputModel:
      type: object
      properties:
        channels:
          type: array
          items:
            $ref: '#/components/schemas/ChannelOutputModel'
        pagination:
          $ref: '#/components/schemas/PaginationOutputModel'
    MessageCreationInputModel:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    MessageOutputModel:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        author:
          type: string
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
    MessageUpdateOutputModel:
      type: object
      properties:
        editedAt:
          type: string
          format: date-time
    MessagesPaginatedOutputModel:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/MessageOutputModel'
        pagination:
          $ref: '#/components/schemas/PaginationOutputModel'
...
    Problem:
      description: Error response
      type: object
      properties:
        type:
          type: string
        title:
          type: string
          format: uri
        status:
          type: integer
        detail:
          type: string
    InputValidationProblem:
      description: Validation error for Body, Query, or Path parameters
      type: object
      properties:
        type:
          type: string
        title:
          type: string
          format: uri
        status:
          type: integer
        detail:
          type: string
        errors:
          type: array
          items:
            type: string
...
    MessageCreationOutputModel:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/Problem'
              - $ref: '#/components/schemas/InputValidationProblem'
    Unauthorized:
        description: Unauthorized
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Problem'
    Forbidden:
        description: Forbidden
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Problem'
    TooManyRequests:
        description: Too many requests
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Problem'
    NotFound:
        description: Not found
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Problem'
...
  parameters:
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 50
        maximum: 100
        minimum: 1
    Name:
      name: name
      in: query
      required: false
      schema:
        type: string
    GetCount:
      name: getCount
      in: query
      required: false
      schema:
        type: boolean
        default: true
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        enum: [ asc, desc ]
        default: asc
    UserID:
      name: userID
      in: path
      required: true
      schema:
        type: number
    ChannelID:
      name: channelID
      in: path
      required: true
      schema:
        type: number
    MessageID:
      name: messageID
      in: path
      required: true
      schema:
        type: number
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID